requirements.txt
arrow==0.14.2
beautifulsoup4==4.7.1
binaryornot==0.4.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
Django==2.2.2
django-bootstrap3==11.0.0
django-modeltranslation==0.13.1
django-vanilla-views==1.0.6
django-webtest==1.9.6
future==0.17.1
idna==2.8
Jinja2==2.10.1
jinja2-time==0.2.0
MarkupSafe==1.1.1
model-mommy==1.6.0
pkg-resources==0.0.0
poyo==0.4.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
urllib3==1.25.3
waitress==1.3.0
WebOb==1.8.5
WebTest==2.0.33
whichcraft==0.5.2
